feat(daemon): re-attach cross-machine memory-pool data plane onto tensor-pool
Port the pre-tensor-pool cross-machine data plane to the upstream
tensor-pool extension seam (dora-rs/dora#3152 rework):

- Daemon holds a TensorPoolManager for the cross_pools table (mirror
  tracking); main pool table stays node-side per the extension model.
- Full data plane restored: zenoh relay (MemoryPoolWrite/RegisterPool/
  RegisterPoolAck/FreePool/MemoryPoolWriteAck), mirror create/write under
  the DORADMA seqlock, direct-TCP listener + persistent connections,
  commit-ack reply withholding with seq matching, per-(dataflow,pool)
  dual-path write locks, orphan mirror sweep, per-dataflow state drain.
- Cross-machine methods moved out of the tensor-pool test module into
  the production impl (they were unreachable behind cfg(test)).
- Auth: DORA_MEMORY_POOL_AUTH_TOKEN enables a token handshake on every
  direct-TCP connection (send + verify sides); rejected peers are
  dropped before any frame touches a mirror, the origin degrades to
  the zenoh relay. Token injected into spawned nodes.
- Opt-in: the whole data plane is gated behind
  DORA_MEMORY_POOL_CROSS_MACHINE=1 (listener, mirroring, register);
  disabled daemons reject cross-machine registration with a clear error.
- Bind config: DORA_MEMORY_POOL_DATA_BIND for the data listener
  (default 0.0.0.0); existing PORT/ADDR envs unchanged.
- FreePinnedMemory daemon request restored as the cross-machine release
  entry (unregister + targeted FreePool + mirror unlink).
- postcard serialization (upstream codec) replaces bincode.

Tests: 13 cross-pool tests (incl. new auth handshake accept/reject) +
full daemon suite (228) green; clippy clean on affected crates.

+++ libraries/message/src/daemon_to_node.rs
@@ -137,7 +137,7 @@ impl DaemonReply {
             // its own length is the whole hint.
             DaemonReply::ExtensionValue { value } => value.as_ref().map_or(0, |bytes| bytes.len()),
             DaemonReply::Result(_) | DaemonReply::NodeConfig { .. } | DaemonReply::Empty => 0,
-            DaemonReply::PinnedMemoryMetadata { metadata } => metadata.parameters.len() as usize,
+            DaemonReply::PinnedMemoryMetadata { metadata } => metadata.parameters.len(),
             DaemonReply::CrossMachinePoolRegistered { .. } => 0,
         }
     }
